+++ demo/nextjs/pages/thumbnail-render-spinner/index.tsx
@@ -1,0 +1,46 @@
+import * as React from 'react';
+import { Viewer, Worker } from '@react-pdf-viewer/core';
+import { thumbnailPlugin } from '@react-pdf-viewer/thumbnail';
+
+const IndexPage = () => {
+    const thumbnailPluginInstance = thumbnailPlugin({
+        renderSpinner: () => <div className="square-spinner" />,
+    });
+    const { Thumbnails } = thumbnailPluginInstance;
+
+    return (
+        <Worker workerUrl="https://unpkg.com/pdfjs-dist@2.10.377/build/pdf.worker.js">
+            <div
+                style={{
+                    border: '1px solid rgba(0, 0, 0, 0.1)',
+                    display: 'flex',
+                    height: '50rem',
+                    margin: '5rem auto',
+                    width: '64rem',
+                }}
+            >
+                <div
+                    style={{
+                        alignItems: 'center',
+                        borderRight: '1px solid rgba(0, 0, 0, 0.1)',
+                        display: 'flex',
+                        padding: '4px',
+                        width: '20%',
+                    }}
+                >
+                    <Thumbnails />
+                </div>
+                <div
+                    style={{
+                        flex: 1,
+                        overflow: 'hidden',
+                    }}
+                >
+                    <Viewer fileUrl="/pdf-open-parameters.pdf" plugins={[thumbnailPluginInstance]} />
+                </div>
+            </div>
+        </Worker>
+    );
+};
+
+export default IndexPage;