+++ packages/thumbnail/src/index.d.ts
@@ -25,6 +25,9 @@ export type RenderCurrentPageLabel = (props: RenderCurrentPageLabelProps) => Rea
 
 export interface ThumbnailPluginProps {
     renderCurrentPageLabel?: RenderCurrentPageLabel;
+    // The spinner that replaces the default `Spinner` component
+    // For example, it is displayed when loading the cover or thumbnail of a page
+    renderSpinner?: () => React.ReactElement;
 }
 
 export interface ThumbnailPlugin extends Plugin {